added dailyrecord model
